requirements.txt
langchain
langchain-core
langchain-community
langchain-google-genai
google-generativeai
python-dotenv
mysql-connector-python
